+++ R/RcppExports.R
@@ -69,7 +69,7 @@ carbon_sapwoodStarchCapacity <- function(SA, H, L, V, woodDensity, conduit2sapwo
     .Call(`_medfate_sapwoodStarchCapacity`, SA, H, L, V, woodDensity, conduit2sapwood)
 }
 
-carbon_carbonCompartments <- function(x, biomassUnits = "kg_m2") {
+carbon_carbonCompartments <- function(x, biomassUnits = "g_m2") {
     .Call(`_medfate_carbonCompartments`, x, biomassUnits)
 }